Remove lecturer info when Coodinator selected

+++ outlineUI.py
@@ -559,6 +559,7 @@ class Root(Tk):
         #values = docx.Document(gv.originalDoc).paragraphs
         next_ = afterNext = None
         isCourseDuration = False
+        isDeleted = False
         length = len(values) 
         valuesFromCD = [gv.courseCode, gv.courseTitle, gv.prerequisites, gv.corequisites, gv.restrictions, gv.nzqfLevel, gv.credits, gv.courseAims, gv.learningOutcomes]
         
@@ -651,10 +652,19 @@ class Root(Tk):
                 else:
                     delete_paragraph(par)
             
-            if "Course Coordinator" in par.text:
-                if self.rbCoursePerson.get() != "Course Coordinator":
+            if self.rbCoursePerson.get() == "Lecturer":
+                if "Course Coordinator" in par.text and isDeleted == False:
                     for i in range(6):
                         delete_paragraph(values[index+i])
+                    isDeleted = True
+            
+            if self.rbCoursePerson.get() == "Course Coordinator":
+                if "+Lecturer Name+" in par.text and isDeleted == False:
+                    for i in range(6):
+                        delete_paragraph(values[index+i])
+                    isDeleted = True
+            
+            
             
             def delete_paragraph(paragraph):
                 p = paragraph._element
